@@ -1,24 +1,29 @@
 buildscript {
-    repositories {
-        jcenter()
-    }
+    if (project == rootProject) {
+        repositories {
+            google()
+            jcenter()
+        }
 
-    dependencies {
-        classpath 'com.android.tools.build:gradle:2.2.3'
+        dependencies {
+            classpath 'com.android.tools.build:gradle:3.4.2'
+        }
     }
 }
 
 apply plugin: 'com.android.library'
 
+def getExtWithFallback(name, fallback) {
+    rootProject.ext.has(name)? rootProject.ext.get(name) : fallback
+}
+
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.3"
+    compileSdkVersion getExtWithFallback('compileSdkVersion', 28)
+    buildToolsVersion getExtWithFallback('buildToolsVersion', '28.0.3')
 
     defaultConfig {
-        minSdkVersion 16
-        targetSdkVersion 22
-        versionCode 1
-        versionName "1.0"
+        minSdkVersion getExtWithFallback('minSdkVersion', 16)
+        targetSdkVersion getExtWithFallback('targetSdkVersion', 28)
     }
     lintOptions {
         abortOnError false
@@ -34,6 +39,6 @@ repositories {
 
 dependencies {
     //noinspection GradleDynamicVersion
-    compile 'com.facebook.react:react-native:+'
-    compile 'com.android.support:exifinterface:25.4.0'
+    implementation 'com.facebook.react:react-native:+'
+    implementation 'androidx.exifinterface:exifinterface:1.0.0'
 }
