@@ -1,7 +1,4 @@
-import {
-  NativeModules,
-  Platform,
-} from 'react-native';
+import { NativeModules, Platform } from 'react-native';
 
 const SUPPORTED_FORMATS = Platform.select({
   ios: ['JPEG', 'PNG'],
@@ -10,7 +7,7 @@ const SUPPORTED_FORMATS = Platform.select({
 
 export default {
   createResizedImage: (path, width, height, format, quality, rotation = 0, outputPath) => {
-    if (! SUPPORTED_FORMATS.includes(format)) {
+    if (!SUPPORTED_FORMATS.includes(format)) {
       throw new Error(`Only ${SUPPORTED_FORMATS.join()} formats are supported by createResizedImage`);
     }
 
